requirements/requirement.txt
django==4.2.2
psycopg2==2.9.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python-dotenv==1.0.0
web3==1.9.2
pyotp==2.8.0
django-cors-headers==4.2.0
django-crispy-forms
dry_rest_permissions
cryptography